requirements.txt
Django==1.11.1
djangorestframework==3.6.3
Markdown==2.6.8
django-filter==1.0.2
#MySQL-python==1.2.3
mysqlclient==1.3.12
moment==0.6.1
xlwt==1.2.0
xlrd==1.2.0
xlutils==2.0.0
pandas==0.20.3
gunicorn==19.7.1
user-agents==1.1.0
django-cors-headers==2.4.0
captcha==0.3
schedule==0.6.0
django-celery==3.2.2
redis==2.10.6
django-extensions==2.1.6